+++ VictoryCenter/VictoryCenter.BLL/Queries/TeamMembers/GetByFilters/GetTeamMembersByFiltersQuery.cs
@@ -2,7 +2,7 @@
 using MediatR;
 using VictoryCenter.BLL.DTOs.TeamMember;
 
-namespace VictoryCenter.BLL.Queries.TeamMembers;
+namespace VictoryCenter.BLL.Queries.TeamMembers.GetByFilters;
 
 public record GetTeamMembersByFiltersQuery(TeamMembersFilterDto TeamMembersFilter)
     : IRequest<Result<List<TeamMemberDto>>>;
