+++ VictoryCenter/VictoryCenter.BLL/Queries/TeamMembers/GetById/GetTeamMemberByIdQuery.cs
@@ -2,7 +2,7 @@
 using MediatR;
 using VictoryCenter.BLL.DTOs.TeamMember;
 
-namespace VictoryCenter.BLL.Queries.TeamMembers;
+namespace VictoryCenter.BLL.Queries.TeamMembers.GetById;
 
-public record GetTeamMemberByIdQuery(long Id) 
+public record GetTeamMemberByIdQuery(long Id)
     : IRequest<Result<TeamMemberDto>>;
