+++ VictoryCenter/VictoryCenter.IntegrationTests/ControllerTests/TeamMembers/GetTeamMemberById/GetTeamMemberById.cs
@@ -9,15 +9,15 @@ using VictoryCenter.BLL.DTOs.Test;
 using VictoryCenter.DAL.Data;
 using VictoryCenter.IntegrationTests.Utils;
 
-namespace VictoryCenter.IntegrationTests.ControllerTests;
+namespace VictoryCenter.IntegrationTests.ControllerTests.TeamMembers.GetTeamMemberById;
 
-public class TeamMembersControllerTests : IClassFixture<VictoryCenterWebApplicationFactory<Program>>
+public class GetTeamMemberById : IClassFixture<VictoryCenterWebApplicationFactory<Program>>
 {
     private readonly HttpClient _client;
     private readonly IServiceScope _scope;
     private readonly VictoryCenterDbContext _dbContext;
 
-    public TeamMembersControllerTests(VictoryCenterWebApplicationFactory<Program> factory)
+    public GetTeamMemberById(VictoryCenterWebApplicationFactory<Program> factory)
     {
         _client = factory.CreateClient();
         _scope = factory.Services.CreateScope();
@@ -25,30 +25,13 @@ public class TeamMembersControllerTests : IClassFixture<VictoryCenterWebApplicat
     }
 
     [Fact]
-    public async Task GetTeamMembers_ShouldReturnOk()
-    {
-        var response = await _client.GetAsync("api/TeamMembers/GetTeamMembers");
-        var responseString = await response.Content.ReadAsStringAsync();
-        
-        var options = new JsonSerializerOptions
-        {
-            PropertyNameCaseInsensitive = true
-        };
-        var responseContent = JsonSerializer.Deserialize<List<TeamMemberDto>>(responseString, options);
-        
-        Assert.True(response.IsSuccessStatusCode);
-        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-        Assert.NotNull(responseContent);
-    }
-
-    [Fact]
     public async Task GetTestDataById_ShouldReturnOk()
     {
         var existingEntity = await _dbContext.TeamMembers.FirstOrDefaultAsync();
-        
+
         var response = await _client.GetAsync($"api/TeamMembers/GetTeamMemberById/{existingEntity!.Id}");
         var responseString = await response.Content.ReadAsStringAsync();
-        
+
         var options = new JsonSerializerOptions
         {
             PropertyNameCaseInsensitive = true
@@ -64,7 +47,7 @@ public class TeamMembersControllerTests : IClassFixture<VictoryCenterWebApplicat
     public async Task GetTestDataById_ShouldFail_NotFound()
     {
         var response = await _client.GetAsync($"api/TeamMembers/GetTeamMemberById/{-1}");
-        
+
         Assert.False(response.IsSuccessStatusCode);
         Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
     }
